pyproject.toml
[project]
name = "monet-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Argenis Vargas",email = "argenis1818@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]


[tool.ruff]
# --- General configuration ---
line-length = 79
target-version = "py312"
exclude = [
    "tests",
    "migrations",
    "__pycache__",
    ".git",
    ".pytest_cache",
    ".venv",
]

[tool.ruff.lint]
select = ["N", "B", "C", "E", "F", "S", "W", "B9"]

# --- Ignored rules: E501 (Line too long) ELIMINADO para dejar que Black haga su trabajo ---
ignore = [
    "S101",  # Use of assert detected
    "E203",  # Whitespace before ':' (black compatibility)
    "E266",  # Too many leading '#' for block comment
    "E402",  # Module level import not at top of file (Ajustar si es posible)
    "E731",  # Do not assign a lambda expression
    "B011",  # Do not call assert False
    "B014",  # Redundant exception
    "N812",  # Lowercase imported as non-lowercase
    "N804",  # Invalid class name
    "C901",  # Function too complex (Recomendación: Refactorizar código en lugar de ignorar)
    "B008",
]

per-file-ignores = { "config/settings/*" = ["F405"], "__init__.py" = ["F401"] }

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

[tool.black]
line-length = 79
target-version = ["py312"]
exclude = '''
/(
  \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | migrations
  | __pycache__
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
include_trailing_comma = true
multi_line_output = 3
line_length = 79
skip = ["migrations", "__init__.py", "manage.py"]
known_first_party = ["src"]
known_third_party = ["pytest", "toml", "django", "rest_framework"]
import_heading_stdlib = "Standard Libraries"
import_heading_future = "Future Libraries"
import_heading_django = "Django Libraries"
import_heading_thirdparty = "Third-party Libraries"
import_heading_localfolder = "Local Folders Libraries"
import_heading_firstparty = "Own Libraries"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true
exclude = ["migrations", "tests"]

[[tool.mypy.overrides]]
module = "src.monet"
disable_error_codes = ["var-annotated"]


[tool.django-stubs]
django_settings_module = "config.settings.base"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2,<9.0.0"
pytest-django = ">=4.11.1,<5.0.0"
flake8 = ">=7.3.0,<8.0.0"
isort = ">=7.0.0,<8.0.0"
black = ">=25.9.0,<26.0.0"
ruff = ">=0.14.0,<0.15.0"
mypy = ">=1.18.2,<2.0.0"
django-stubs = ">=5.2.7,<6.0.0"

[tool.poetry]
package-mode = false
